Factor Expr stuff into a private module
